large change

+++ SCNN_keras.py
@@ -18,10 +18,10 @@ def negative_log_likelihood(E):
 		log_risk = K.log(K.cumsum(hazard_ratio))
 		uncensored_likelihood = K.transpose(y_pred) - log_risk
 		censored_likelihood = uncensored_likelihood * E
-		num_observed_event = K.sum([int(e) for e in E]) or 1
+		num_observed_event = K.sum([float(e) for e in E]) or 1
 		return K.sum(censored_likelihood) / num_observed_event * (-1)
 	return loss
-    
+
 """SCNN model, a much small structure of PNS"""
 # complex
 def gen_model():
